Pushback
I now have an optional route, and handle queuing when the way isn't clear.
This involves a third signal which drives the "resolve" phase.

In this phase, we look at our state and decide that we hate it.
Then we call rollback on the people who sent us this garbage.