Added Play Controls
I now have buttons.
They pause and play, and when paused step the action.

The reset button current doesn't work.

I had to change to count the frame number in game.
It used to be the case that the timer would output an incrementing number.
Now, though, when I step that number doesn't exist, and when I pause and play
it goes back to the start.

So, I have to count for myself now.

I couldn't figure out a good way to count in signals without it jumping by 15
every time because of all the listeners...

Dunno.
This is fine for now.