Change to Editable Grid Layout
I now build a grid of empty cells, and have the beginnings of an editable grid
working.

When I click on cells it builds a new cell a populates it out of the current
selection from the signals list.

I've verified that the boxes still move, given that I can only make horizontal
things for now.

I've added a grid so that the pieces aren't just hanging out in the middle of
nowhere.