Fix Step for New Cells
So, this was a weird one...
The issue was when I added a new cell, "Step" didn't work, but play did...

The new cells just didn't get the update signal, but they did get the resolve
and commit signals.
But once I'd played once, step started working again.
It was super weird.

Turns out the issue was my update signal wasn't being shared, so the new cells
were waiting for their own update to the play pause state, independently of all
of the existing cells.

So they didn't step.
Once I pushed play, all of the existing ones switched over, and all of the new
ones switched for the first time.
From then on they all followed together.

Very confusing.
Probably I want all signals to be shared, since I specifically want it to be a
broadcast system.
I'll consider that later (like the next time something weird bites me in the
face).