Fix Task Progress Staying at 10
When I refactored, I didn't notice this.

Turns out that if you fire something that causes pipeline changes inside a
doOnNext weird stuff happens...

Basically, it fired the whole reset chain and everything like you'd expect, but
then afterwards it continued processing.

This meant it got the reset, and wrote 0 in the box as expected, but because it
was a doOnNext it then continued down the chain and wrote 10 into the box.
Poo.

Making it two subscriptions is more right anyway, but it also fixed this issue
because now the chains are separate and when one makes a change, it makes the
other get the right value the first time it runs.