Do "Actual" Data Binding
I had a couple places now that were just taking a signal and putting the value
into a text element.

So, now I just have a method for that.
I just say "This signal should print here".

It's also slightly more efficient, because now I only query the document once,
and then I hold that reference in the function, which is good.