Fancy "You Did It" screen
While this is fancy, it's also working around an issue with timing I was
having.

The level was being completed in the course of 1 tile's update, which meant
that it was processing the success and doing the reset, and then following it
up with continuing to commit the lifecycle events.

This was leading some items to not appear reset.

But, it's a thing I wanted to play with anyway.

I'm not sure how I feel about the way I'm handling things.
I think I dislike it, actually, but I'll think about what I've done.

The resource requirements seem to have gone up quite a bit somehow, but I think
that was just when the inspector was open.
If it becomes an issue, I'll have to look into it closer.

Closes #13