Added zipLatest
There was a situation where we'd get behind due to processing lag.
Then when we'd hit pause, it'd progress a number of frames anyway.

This is because I was using "zip" to combine the two signals, and zip queues,
so when things get behind it just waits.

Now I throw out clock ticks that happen while we're processing, but unlike
"sample", I don't throw out a clock tick just because it missed it, it just
spits it out when the commit is finished.
It just makes sure it doesn't spit out more than one, so they don't back up.